some changes to the admin module
