+++ src/app/component/admin/components/confirm-modal/confirm-modal.component.html
