+++ src/app/component/admin/components/index.ts
@@ -1,0 +1,8 @@
+export { AdminNavComponent } from './admin-nav/admin-nav.component';
+export { ConfirmModalComponent } from './confirm-modal/confirm-modal.component';
+export { ErrorComponent } from './error/error.component';
+export { FeedbacksComponent } from './feedbacks/feedbacks.component';
+export { DialogPhotoComponent } from './feedbacks/dialog-photo/dialog-photo.component';
+export { PlacesComponent } from './places/places.component';
+export { UpdateCafeComponent } from './update-cafe/update-cafe.component';
+export { UsersComponent } from './users/users.component';
