+++ src/app/component/admin/components/places/places.component.css
@@ -1,4 +1,4 @@
-@import '~@angular/material/prebuilt-themes/deeppurple-amber.css';
+@import '../../../../../../node_modules/@angular/material/prebuilt-themes/deeppurple-amber.css';
 
 table {
   width: 100%;
