+++ src/app/component/admin/components/update-cafe/update-cafe.component.css
@@ -1,4 +1,4 @@
-@import "~@ng-select/ng-select/themes/default.theme.css";
+@import "../../../../../../node_modules/@ng-select/ng-select/themes/default.theme.css";
 
 #weekDays-selector input[type=checkbox] {
   display: none!important;
