+++ src/app/component/admin/components/update-cafe/update-cafe.component.html
@@ -1,4 +1,4 @@
-<script src="../../../model/pageable-dto.model.ts"></script>
+<script src="../../../../model/pageable-dto.model.ts"></script>
 <h2 mat-dialog-title style="text-align: center; font-weight: 400;">{{'update-cafe.edit-cafe' | translate}}</h2>
 <hr>
 <mat-dialog-content>
