+++ src/app/component/admin/models/comment-admin-dto.model.ts
@@ -1,5 +1,5 @@
-import {AdminPlace} from '../place/admin-place.model';
-import {PhotoAdminModel} from '../photo-admin.model';
+import { AdminPlace } from './admin-place.model';
+import { PhotoAdminModel } from './photo-admin.model';
 
 export class CommentAdminDto {
   id: number;
